+++ rust/lance-core/src/io.rs
@@ -19,8 +19,10 @@ use std::ops::{Range, RangeFrom, RangeFull, RangeTo};
 
 use arrow_array::UInt32Array;
 
+pub mod commit;
 pub mod local;
 pub mod object_reader;
+pub mod object_store;
 pub mod object_writer;
 mod stream;
 mod traits;
