+++ rust/lance/src/io/commit.rs
@@ -33,38 +33,33 @@
 //! terms of a lock. The trait [CommitLock] can be implemented as a simpler
 //! alternative to [CommitHandler].
 
+use std::fmt::Debug;
 use std::future;
 use std::sync::Arc;
-use std::{fmt::Debug, sync::atomic::AtomicBool};
 
 #[cfg(feature = "dynamodb")]
 pub(crate) mod dynamodb;
 pub(crate) mod external_manifest;
 
+use futures::future::Either;
+use futures::stream::BoxStream;
+use futures::{StreamExt, TryStreamExt};
+use lance_core::{
+    format::{pb, Index, Manifest},
+    io::commit::{CommitError, CommitHandler, ManifestWriter},
+    Error, Result,
+};
+use object_store::path::Path;
+use prost::Message;
+use snafu::{location, Location};
+
+use super::deletion::read_deletion_file;
+use super::ObjectStore;
 use crate::dataset::fragment::FileFragment;
 use crate::dataset::transaction::{Operation, Transaction};
 use crate::dataset::{write_manifest_file, ManifestWriteConfig};
 use crate::format::{DeletionFile, Fragment};
-use crate::Result;
-use crate::{format::pb, format::Index, format::Manifest};
-use crate::{Dataset, Error};
-use futures::future::{BoxFuture, Either};
-use futures::stream::BoxStream;
-use futures::{StreamExt, TryStreamExt};
-use object_store::path::Path;
-use object_store::Error as ObjectStoreError;
-use prost::Message;
-
-use super::deletion::read_deletion_file;
-use super::ObjectStore;
-use snafu::{location, Location};
-/// Function that writes the manifest to the object store.
-pub type ManifestWriter = for<'a> fn(
-    object_store: &'a ObjectStore,
-    manifest: &'a mut Manifest,
-    indices: Option<Vec<Index>>,
-    path: &'a Path,
-) -> BoxFuture<'a, Result<()>>;
+use crate::Dataset;
 
 const LATEST_MANIFEST_NAME: &str = "_latest.manifest";
 const VERSIONS_DIR: &str = "_versions";
@@ -162,305 +157,6 @@ async fn write_latest_manifest(
         .rename(&staging_path, &latest_path)
         .await?;
     Ok(())
-}
-
-fn parse_version_from_path(path: &Path) -> Result<u64> {
-    path.filename()
-        .and_then(|name| name.split_once('.'))
-        .filter(|(_, extension)| *extension == "manifest")
-        .and_then(|(version, _)| version.parse::<u64>().ok())
-        .ok_or(crate::Error::Internal {
-            message: format!("Expected manifest file, but found {}", path),
-        })
-}
-
-/// Handle commits that prevent conflicting writes.
-///
-/// Commit implementations ensure that if there are multiple concurrent writers
-/// attempting to write the next version of a table, only one will win. In order
-/// to work, all writers must use the same commit handler type.
-/// This trait is also responsible for resolving where the manifests live.
-#[async_trait::async_trait]
-pub(crate) trait CommitHandler: Debug + Send + Sync {
-    /// Get the path to the latest version manifest of a dataset at the base_path
-    async fn resolve_latest_version(
-        &self,
-        base_path: &Path,
-        object_store: &ObjectStore,
-    ) -> std::result::Result<Path, crate::Error> {
-        // TODO: we need to pade 0's to the version number on the manifest file path
-        Ok(current_manifest_path(object_store, base_path).await?)
-    }
-
-    // for default implementation, parse the version from the path
-    async fn resolve_latest_version_id(
-        &self,
-        base_path: &Path,
-        object_store: &ObjectStore,
-    ) -> std::result::Result<u64, crate::Error> {
-        let path = self.resolve_latest_version(base_path, object_store).await?;
-
-        parse_version_from_path(&path)
-    }
-
-    /// Get the path to a specific versioned manifest of a dataset at the base_path
-    async fn resolve_version(
-        &self,
-        base_path: &Path,
-        version: u64,
-        _object_store: &ObjectStore,
-    ) -> std::result::Result<Path, crate::Error> {
-        Ok(manifest_path(base_path, version))
-    }
-
-    /// List manifests that are available for a dataset at the base_path
-    async fn list_manifests<'a>(
-        &self,
-        base_path: &Path,
-        object_store: &'a ObjectStore,
-    ) -> Result<BoxStream<'a, Result<Path>>> {
-        list_manifests(base_path, object_store).await
-    }
-
-    /// Commit a manifest.
-    ///
-    /// This function should return an [CommitError::CommitConflict] if another
-    /// transaction has already been committed to the path.
-    async fn commit(
-        &self,
-        manifest: &mut Manifest,
-        indices: Option<Vec<Index>>,
-        base_path: &Path,
-        object_store: &ObjectStore,
-        manifest_writer: ManifestWriter,
-    ) -> std::result::Result<(), CommitError>;
-}
-
-/// Errors that can occur when committing a manifest.
-#[derive(Debug)]
-pub enum CommitError {
-    /// Another transaction has already been written to the path
-    CommitConflict,
-    /// Something else went wrong
-    OtherError(crate::Error),
-}
-
-impl From<crate::Error> for CommitError {
-    fn from(e: crate::Error) -> Self {
-        Self::OtherError(e)
-    }
-}
-
-impl From<CommitError> for crate::Error {
-    fn from(e: CommitError) -> Self {
-        match e {
-            CommitError::CommitConflict => Self::Internal {
-                message: "Commit conflict".to_string(),
-            },
-            CommitError::OtherError(e) => e,
-        }
-    }
-}
-
-/// Whether we have issued a warning about using the unsafe commit handler.
-static WARNED_ON_UNSAFE_COMMIT: AtomicBool = AtomicBool::new(false);
-
-/// A naive commit implementation that does not prevent conflicting writes.
-///
-/// This will log a warning the first time it is used.
-pub struct UnsafeCommitHandler;
-
-#[async_trait::async_trait]
-impl CommitHandler for UnsafeCommitHandler {
-    async fn commit(
-        &self,
-        manifest: &mut Manifest,
-        indices: Option<Vec<Index>>,
-        base_path: &Path,
-        object_store: &ObjectStore,
-        manifest_writer: ManifestWriter,
-    ) -> std::result::Result<(), CommitError> {
-        // Log a one-time warning
-        if !WARNED_ON_UNSAFE_COMMIT.load(std::sync::atomic::Ordering::Relaxed) {
-            WARNED_ON_UNSAFE_COMMIT.store(true, std::sync::atomic::Ordering::Relaxed);
-            log::warn!(
-                "Using unsafe commit handler. Concurrent writes may result in data loss. \
-                 Consider providing a commit handler that prevents conflicting writes."
-            );
-        }
-
-        let version_path = self
-            .resolve_version(base_path, manifest.version, object_store)
-            .await?;
-        // Write the manifest naively
-        manifest_writer(object_store, manifest, indices, &version_path).await?;
-
-        write_latest_manifest(&version_path, base_path, object_store).await?;
-
-        Ok(())
-    }
-}
-
-impl Debug for UnsafeCommitHandler {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("UnsafeCommitHandler").finish()
-    }
-}
-
-/// A commit implementation that uses a temporary path and renames the object.
-///
-/// This only works for object stores that support atomic rename if not exist.
-pub struct RenameCommitHandler;
-
-#[async_trait::async_trait]
-impl CommitHandler for RenameCommitHandler {
-    async fn commit(
-        &self,
-        manifest: &mut Manifest,
-        indices: Option<Vec<Index>>,
-        base_path: &Path,
-        object_store: &ObjectStore,
-        manifest_writer: ManifestWriter,
-    ) -> std::result::Result<(), CommitError> {
-        // Create a temporary object, then use `rename_if_not_exists` to commit.
-        // If failed, clean up the temporary object.
-
-        let path = self
-            .resolve_version(base_path, manifest.version, object_store)
-            .await?;
-
-        // Add .tmp_ prefix to the path
-        let mut parts: Vec<_> = path.parts().collect();
-        // Add a UUID to the end of the filename to avoid conflicts
-        let uuid = uuid::Uuid::new_v4();
-        let new_name = format!(
-            ".tmp_{}_{}",
-            parts.last().unwrap().as_ref(),
-            uuid.as_hyphenated()
-        );
-        let _ = std::mem::replace(parts.last_mut().unwrap(), new_name.into());
-        let tmp_path: Path = parts.into_iter().collect();
-
-        // Write the manifest to the temporary path
-        manifest_writer(object_store, manifest, indices, &tmp_path).await?;
-
-        let res = match object_store
-            .inner
-            .rename_if_not_exists(&tmp_path, &path)
-            .await
-        {
-            Ok(_) => Ok(()),
-            Err(ObjectStoreError::AlreadyExists { .. }) => {
-                // Another transaction has already been committed
-                // Attempt to clean up temporary object, but ignore errors if we can't
-                let _ = object_store.inner.delete(&tmp_path).await;
-
-                return Err(CommitError::CommitConflict);
-            }
-            Err(e) => {
-                // Something else went wrong
-                return Err(CommitError::OtherError(e.into()));
-            }
-        };
-
-        write_latest_manifest(&path, base_path, object_store).await?;
-
-        res
-    }
-}
-
-impl Debug for RenameCommitHandler {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("RenameCommitHandler").finish()
-    }
-}
-
-/// A commit implementation that uses a lock to prevent conflicting writes.
-#[async_trait::async_trait]
-pub trait CommitLock: Debug {
-    type Lease: CommitLease;
-
-    /// Attempt to lock the table for the given version.
-    ///
-    /// If it is already locked by another transaction, wait until it is unlocked.
-    /// Once it is unlocked, return [CommitError::CommitConflict] if the version
-    /// has already been committed. Otherwise, return the lock.
-    ///
-    /// To prevent poisoned locks, it's recommended to set a timeout on the lock
-    /// of at least 30 seconds.
-    ///
-    /// It is not required that the lock tracks the version. It is provided in
-    /// case the locking is handled by a catalog service that needs to know the
-    /// current version of the table.
-    async fn lock(&self, version: u64) -> std::result::Result<Self::Lease, CommitError>;
-}
-
-#[async_trait::async_trait]
-pub trait CommitLease: Send + Sync {
-    /// Return the lease, indicating whether the commit was successful.
-    async fn release(&self, success: bool) -> std::result::Result<(), CommitError>;
-}
-
-#[async_trait::async_trait]
-impl<T: CommitLock + Send + Sync> CommitHandler for T {
-    async fn commit(
-        &self,
-        manifest: &mut Manifest,
-        indices: Option<Vec<Index>>,
-        base_path: &Path,
-        object_store: &ObjectStore,
-        manifest_writer: ManifestWriter,
-    ) -> std::result::Result<(), CommitError> {
-        let path = self
-            .resolve_version(base_path, manifest.version, object_store)
-            .await?;
-        // NOTE: once we have the lease we cannot use ? to return errors, since
-        // we must release the lease before returning.
-        let lease = self.lock(manifest.version).await?;
-
-        // Head the location and make sure it's not already committed
-        match object_store.inner.head(&path).await {
-            Ok(_) => {
-                // The path already exists, so it's already committed
-                // Release the lock
-                lease.release(false).await?;
-
-                return Err(CommitError::CommitConflict);
-            }
-            Err(ObjectStoreError::NotFound { .. }) => {}
-            Err(e) => {
-                // Something else went wrong
-                // Release the lock
-                lease.release(false).await?;
-
-                return Err(CommitError::OtherError(e.into()));
-            }
-        }
-        let res = manifest_writer(object_store, manifest, indices, &path).await;
-
-        write_latest_manifest(&path, base_path, object_store).await?;
-
-        // Release the lock
-        lease.release(res.is_ok()).await?;
-
-        res.map_err(|err| err.into())
-    }
-}
-
-#[async_trait::async_trait]
-impl<T: CommitLock + Send + Sync> CommitHandler for Arc<T> {
-    async fn commit(
-        &self,
-        manifest: &mut Manifest,
-        indices: Option<Vec<Index>>,
-        base_path: &Path,
-        object_store: &ObjectStore,
-        manifest_writer: ManifestWriter,
-    ) -> std::result::Result<(), CommitError> {
-        self.as_ref()
-            .commit(manifest, indices, base_path, object_store, manifest_writer)
-            .await
-    }
 }
 
 #[derive(Debug, Clone)]
